update: complete dropdown rework, discount + tax handling
Dropdown reworked to reside inside each items card. Complete logic in place to add/remove Discounts and Taxes.

Wrapper component made to pass the tanstack client to the CartContainer. Need to implement the handleCalculateOrder() method in CartContainer and display in the Grand Total Card. Also need to make the GT Card more verbose to hold the Discount and Tax Amount.