update: responsive images added, close cart button and search query fix
Images once previously not shown for tablet or phone screens is now being shown for the both of them. Close Cart button not visible on phone made closing the cart impossible. Fixed it for smaller screens. And finally fixed a bug with the Search where when having no category, it wouldn't properly search the textFilter