update: add multiple custom hooks + big fixes
Created multiple custom hooks. useCartMutation, useFetchDiscounts, useFetchTaxes.

Had an issue where Order would not be calculated if a 'Free' element was in the cart. Turns out the Free element was actually 'Variable' price which means it would not have a fixed one. So we cant calculate for it. Handled this in the UI where Variable Cost items cannot be added to the Cart.

Changed the order calculation logic to auto calculate order and not on the basis of a button event.